+++ pip_audit/_dependency_source/__init__.py
@@ -2,7 +2,13 @@
 Dependency source interfaces and implementations for `pip-audit`.
 """
 
-from .interface import PYPI_URL, DependencyFixError, DependencySource, DependencySourceError
+from .interface import (
+    PYPI_URL,
+    DependencyFixError,
+    DependencySource,
+    DependencySourceError,
+    InvalidRequirementSpecifier,
+)
 from .pip import PipSource, PipSourceError
 from .pyproject import PyProjectSource
 from .requirement import RequirementSource
@@ -12,6 +18,7 @@ __all__ = [
     "DependencyFixError",
     "DependencySource",
     "DependencySourceError",
+    "InvalidRequirementSpecifier",
     "PipSource",
     "PipSourceError",
     "PyProjectSource",